requirements.txt
apache-airflow-providers-amazon==10.2.1
boto3==1.28.57
botocore==1.31.57
